+++ azure-storage-common/ChangeLog.md
@@ -7,4 +7,5 @@
 * Moved method `SharedAccessSignatureHelper::generateFileServiceSharedAccessSignatureToken()` into `FileSharedAccessSignatureHelper`.
 * `CommonMiddleWare` constructor requires storage service version as parameter now.
 * `AccessPolicy` class is now an abstract class, added children classes `BlobAccessPolicy`, `ContainerAccessPolicy`, `TableAccessPolicy`, `QueueAccessPolicy`, `FileAccessPolicy` and `ShareAccessPolicy`.
+* Fixed a bug that `Utilities::allZero()` will return true for non-zero data chunks.
 * Deprecated PHP 5.5 support.